refactor: Main.jsx App.js에서 분리
pitcher_windup.png 크기조정, Main.jsx 생성에 따른 import 경로 수정
